doc strings + renaming
